Adding option ${PROJECT_NAME}_ELEVATE_SS_TO_PS (VRI Kanban #2951)
A TriBITS project (or a user) can now set the option
${PROJECT_NAME}_ELEVATE_SS_TO_PS=ON and every listed SS package or subpackage
will be elevated to PS.  I added a basic unit test that shows this works and I
manually valiated it on VERA.

There is a long history but basically, differenting between PS and SS packages
is not useful for VERA.  The reason that non-Trilinos VERA packages were
labeled as SS was to allow testing both Trilinos and non-Trilinos packages
prior to pushing.  This is no longer possible because VERA works with a clone
of Trilinos (synced very frequently).  Therefore, rather than require all the
various VERA repos to update from SS to PS, I just added an option to do it
automatically.  That way, we can now use the checkin-test.py --default-builds
for VERA builds and stop using the --ss-extra-builds (which was very
confusing for VERA developers).

Note that it is still very useful for Trilinos (and perhaps other projects) to
differentiate PS and SS packages and code and therefore the given behavior
will be maintained in the checkin-test.py script.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=109,notpassed=0 (1.99 min)
1) SERIAL_RELEASE => passed: passed=109,notpassed=0 (2.25 min)
2) MPI_DEBUG_SS => passed: passed=109,notpassed=0 (1.94 min)
3) SERIAL_RELEASE_SS => passed: passed=109,notpassed=0 (1.36 min)