Ifpack2: Fixed BlockRelaxation::DoSGS. Moved the LRID push_back in OverlappingPartitioner to the end.
The first change finishes fixing block GS in the serial case. The AdditiveSchwarz problems
still remain, and are still being fixed. I don't know if BlockRelaxation will still behave
correctly in (non-AdditiveSchwarz) parallel.

The second change fixes an issue with Vanka smoothing in particular. The way it originally
was, the pivot entry for the first row of each block would be zero for Vanka smoothing for
the incompressible Stokes equations using stable elements. This way, the pivot should not
be zero.