split IO into two phases to improve efficiency.
The first phase reads mesh entities and connectivity,
the second phase reads coordinates, attributes, other field-data.

This allows us to take advantage of a new stk-mesh optimization where
the allocation of field-data can be delayed until the mesh is fully populated.
The benefit of this is that it avoids copying field-data in and out of the
temporary buckets that get created and deleted while entities are being moved
to new part memberships as connections are being populated, etc. Rather, we
wait until all part memberships are finalized, all entities are in their final
buckets, and then field-data is laid out. The overall result is that the memory
high-water-mark is reduced.

This improvement is so far only done in frio (framework io) so it benefits
fmwk-based apps. The same improvement will be done in stk-io next, so that
nalu will gain the same benefit.

This improvement helps 'complicated' meshes significantly, but has little or no
effect on 'simple' meshes. (A complicated mesh has lots of parts, different
types of elements, etc.) For a 'simple' mesh, nodes don't move in and out of
temporary buckets nearly as much as connectivities get populated, so the
high-water-mark for field-data doesn't change as much.

A couple of examples:
adagio/shell_tri3d_nodal_probe (this test is the original subject of the
user-support ticket reported by Nate Crane)
    high-water-mark goes from 5.5GB to 2.0GB

conchas/kEpsCavityDES
    high-water-mark goes from 87MB to 78MB

adagio/p007_two_layer_cake
    high-water-mark goes from 262MB to 260MB

A few other tests from adagio and conchas that I sampled, showed no significant change in reported memory.

Note: currently, this improvement doesn't benefit restart runs. Restart runs
still use the same IO sequence as before. It will be more difficult to alter
the way restarts are done. A backlog story will be proposed to enable these
improvements for restart runs.

(From Sierra)