Tpetra: CrsGraph: Replaced std::deque with std::vector for nonlocals.
CrsGraph was storing nonlocal data (data in rows that are not owned by
the calling process) as an std::map<GO, std::deque<GO> >.  There was no
reason to use a deque ("double-ended queue"), since the only insertion
operations on it were push_back (at the end of the object).  Thus, I
replaced std::deque with std::vector.