Ifpack2: Fixed a couple of small bugs
Changed the return type of the function numLocalParts() in Partitioner
to be an int instead of LocalOrdinal. The documentation in
Ifpack2_OverlappingPartitioner_decl.hpp explains why this is the
correct type. Also, there was a bug in a for loop in the
UserPartitioner (trying to use ++i instead of ++ii). Fixed.