Ifpack2: Fixed a unit test so that it succeeds now.
The partitioning unit test still wanted "partitioner: local parts" to
be a LocalOrdinal. As explained in Ifpack2_OverlappingPartitioner_decl.hpp,
this type needs to be an int, since this could be a negative number
with special significance. The test will fail if a LocalOrdinal is
passed to the parameter list. Thus it now passes an int to the
parameter list.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=22,notpassed=0 (2.43 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (2.67 min)
Other local commits for this build/test group: c2abd68