Tpetra: CrsMatrix, Map: Improved public documentation and clone().
CrsMatrix's clone() implementation had some code that was hard to
read, because it was not conforming to Tpetra's code standards.
Since summer students are working on Ifpack2 now and need clone()
to work, I wanted to make sure that clone()'s implementation looked
as it should look before they had to look at it.