KokkosClassic,Tpetra: Changed Kokkos namespace to KokkosClassic.
This will make it possible for Kokkos Classic and Kokkos Array to
coexist in the same program.  We will need this to support gradual
porting of Tpetra (and downstream packages) to use Kokkos Array.

For now, Kokkos_ConfigDefs.hpp imports everything in the KokkosClassic
namespace into the Kokkos namespace.  This is a temporary fix that will
go away once downstream packages and software have time to change
namespaces.  Thus, this commit should not break backwards compatbility.

This commit also fixes forward declarations to Kokkos Classic classes in
Tpetra.  The aforementioned fix does not affect forward declarations of
classes in the Kokkos namespace, that do not include Kokkos Classic
header files.  This commit tells forward declarations of Kokkos Classic
classes in Tpetra, to use the KokkosClassic namespace instead of the
Kokkos namespace.