Piro: Fixed handling of parameter nominal values
Used the value of the first parameter (identified by the index 0) from
inArgs if and only if it is non-null, instead of unconditionally using
it. Otherwise, the corresponding nominal value is used by default.

According to the specification of Thyra::ModelEvaluator::evalModel, all
supported parameters that are not set by the client must be assumed to
have the nominal value.

The former implementation of RythmosSolver::evalModel did not fulfill
that contract since it always ignored its own parameter nominal value.