Xpetra: added clone() functionality for Map and Matrix
In many cases, we want to construct some data on CPU and then migrate it
to GPU. The current way of doing that in Tpetra is to use clone()
functionality, which allows the change of Node type. This patch tries to
propagate that functionality into Xpetra, which is going to be exercised
by MueLu.