MueLu: special treatment for xml sublists named "ParameterList"
In many situations, we want to pass the parameters from the xml file to
a third party application. For instance, we may want to pass a list of
parameters to a partitioner, or to a smoother. There was no general
mechanism in place to do that. Now, for factories which follow the new
standard of validated parameter lists, we reserve a sublist with name
"ParameterList" to be passed to applications.