Ifpack2: making AdditiveSchwarz a little clearer
Adding type of subdomain reordering to AdditiveSchwarz description.
Making use of RCM explicit in AdditiveSchwarz.  This doesn't change the behavior, but previously this class relied on the fact that the default in Zoltan2 is
RCM.
Changed "schwarz: reordering type" to "schwarz: use reordering".  This makes sense because the reordering type is actually a Zoltan2 parameter.  Ifpack2 only
cares that reordering occurs -- it's up to the caller to make sure the Zoltan2 params are set correctly.