Tpetra: Added "long long" to the list of GlobalOrdinal ETI types.
If the CMake BOOL option Teuchos_ENABLE_LONG_LONG_INT is ON, then
Tpetra will add "long long" to the list of GlobalOrdinal (and Scalar)
types for which it does explicit template instantiation (ETI).  This
is the fruit of a discussion started earlier this week by Eric Cyr.
We have to add it to the list of Scalar types as well as GlobalOrdinal
types, because some downstream packages use Scalar=GlobalOrdinal for
communication.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=135,notpassed=0 (3.18 min)
3) SERIAL_RELEASE_SS => passed: passed=118,notpassed=0 (0.30 min)
Other local commits for this build/test group: 271bd52