Ifpack2: ILUT: Changed fill level from magnitude_type to int.
Fill level must be an integer, but it is currently exposed in the
interface and the ParameterList as a magnitude_type (double, if
scalar_type is double).  This commit fixes that, so that fill level
is exposed as an int.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=136,notpassed=0 (3.15 min)
3) SERIAL_RELEASE_SS => passed: passed=119,notpassed=0 (0.25 min)