Piro: Added a basic Thyra-based LOCA solver
LOCASolver is a class template parametrized by the scalar type, even if
LOCA only supports double currently. As such, it follows the convention
set by the other Thyra-based Piro solvers, and allows for future
extension later without requiring breaking changes or the introduction
of a new template class. The class template LOCASolver and its companion
factory function are always explicitly instantiated for Scalar = double,
and the definitions are never included in the main header file.

For the time being, the solver only computes the solution and responses.
The sensitivities are not yet supported. Also, only basic uses have
been tested.