Adding missing include after rearrangement of Thyra headers (Trilinos #5902)
If code uses a type Namespace::X, then it needs to directly include the header
Namespace_X.hpp or there is no guarantee that the definition for Namespace::X
will be available.  That is just the way it is and this is not breaking
backward compatibility except in the most rigid sense.  This change would have
worked with a version of Thyra 7 years ago.