Tpetra:  Enable faster multivector import/export for GPU node.
This commit substantially improves the performance of import/export of
multivectors using the KokkosClassic::ThrustGPUNode.  Currently import
and export of ThrustGPUNode multivectors copies the entire multivector
to the host, packs the import/export buffers on the host, unpacks the
communicated buffers on the host, and then copies the new multivector
back to the GPU.  This dramatically degrades performance in e.g., the
import from domain to column map and export from row to range map in
SpMv operations with a cost proportional to the vector size.

This commit fixes this problem by supplying kernels to pack and unpack
the buffers directly on the GPU.  Furthermore, if the MPI library
supports GPU RDMA, the packed bufferes can be communicated without ever
transfering them to the host.

This was implemented by supplying a new version of Tpetra::DistObject
(called DistObjectKA) that uses Kokkos::View to store the pack/unpack
buffers and uses Kokkos::parallel_for the pack/unpack operations.
Furthermore new versions of Distributor::doPosts() et al are provided
using Kokkos::View allowing the buffers to communicated directly via
RDMA or transfered to the host.

This has several implications for dependencies:
  * I put a number of helper functions in the Kokkos::Compat package.
    This package also manages the explicit instantiations necessary so
    that Tpetra doesn't need to be compiled by nvcc.
  * Tpetra now has optional dependencies on Kokkos::Core and
    Kokkos::Compat.
  * KokkosClassic has an optional dependency on Kokkos::Core so that the
    ThrustGPUNode can call Kokkos::Cuda::initialize().  Note that there
    is currently no automatic way of calling Kokkos::Cuda::finalize() as
    there is no reliable way to do this (putting it in the node
    destructor is a bad idea since the user is allowed to instantiate
    multiple nodes).  This is something that will have to be addressed
    as Kokkos::View is integrated throughout Tpetra.

All of this is optional and by default non of it is enabled (thus the
default behavior is expensive GPU import/export).  To turn this on you
need to enable:

 -D Trilinos_ENABLE_Kokkos=ON \
 -D Trilinos_ENABLE_KokkosCore=ON \
 -D Trilinos_ENABLE_KokkosClassic=ON \
 -D Trilinos_ENABLE_KokkosCompat=ON \
 -D Kokkos_ENABLE_Cuda:BOOL=ON \
 -D Tpetra_ENABLE_KokkosCore=ON \
 -D Tpetra_ENABLE_Kokkos_DistObject=ON \

This last option is the one that changes the base class of MultiVector
to DistObjectKA.  Additionally, if you want to enable RDMA, add

 -D Tpetra_ENABLE_MPI_CUDA_RDMA=ON \

Alternatively you can set "Enable MPI CUDA RDMA support" = true|false in
the Distributor parameter list to enable RDMA.  The default is the value
of Tpetra_ENABLE_MPI_CUDA_RDMA which defaults to OFF.

When turning all of this on, all of the Tpetra tests still pass.  The
only problem is Kokkos::Cuda::finalize() isn't called automatically,
creating a small heap leak.  I did not change any of the tests to call
finalize().