Belos: Fixed compile error (2.0 * ScalarType) in BlockFGmresIter.
Thanks to Jonas Thies (Deutsches Zentrum fuer Luft- und Raumfahrt),
who discovered this problem in the 11.2.4 release, for pointing this
out.  operator* (2.0, ScalarType) does not work for all possible
ScalarType types.  In particular, Jonas experienced troubles with
ScalarType = std::complex<float>.  This is a bit surprising to me, but
the problem is easy to fix, so I've done it.  I've also replaced other
instances of double or float literals (0.0 and 1.0) with the
appropriate references to Teuchos::ScalarTraits<ScalarType>::{zero,
one}(), and reformatted the file to make it easier to read.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosClassic, KokkosCompat, Tpetra, Belos, Ifpack2, Xpetra
0) MPI_DEBUG => passed: passed=164,notpassed=0 (3.48 min)
1) SERIAL_RELEASE => passed: passed=145,notpassed=0 (0.33 min)