MueLu: use slow assembly routine in TentativePFactory if there are empty processors
TODO: find a fix to be able to use the expert routines with empty processors.
The error message is
"Teuchos::ArrayRCP<double>::assert_in_range: Error, [lowerOffset,lowerOffset+size-1] = [0,0] does not lie in the range [0,-1]!"
We have to fill the matrix data using the expert routines properly for "empty" procs.