MueLu: Level: prevent hangup in the Clear() function
To prevent infinite looping, we need to check whether we have actually
removed the data. In buggy code it may happen that we were unable to do
that, for instance, if the data was outstanding request. At the moment,
we throw if this occurs, because it is very troubling in reuse case,
where you cannot say anymore what's going on, and which data is being
used.