MueLu: Smoothers: DeclareInput fix and cleanup
We had a potental bug lurking in the DirectSolver. The problem was that
DirectSolver did not call DeclareInput on Amesos[2]Smoother. How could
it, when it hadn't construct the smoother prior to DeclareInput call.
That messed up the dependency tree, leaving coarsest "A" with
outstanding request.

Fixing it turned out not to be easy. I added lib to Hierarchy and Level,
and moved Amesos[2]Smoother constructors to DirectSolver constructor. By
doing that, we are able to propagate DeclareInput.