Adding summary string for top commit (Trilinos #5967)
When working in local branches or with collaboration branches or any other
sitution where the SHA1 of the top commit is not what it will be in the global
repo branch (i.e. origin/master), it is very advantagous to see the summary
message to be able to identify the commit.  Also, that short summary message
can give a lot of other info and is just nice to see.

I limited the output of the summary message to just 80 chars so that this does
not get crazy like some bad commit messages get.