Ifpack2: OverlapGraph: Cleaned up class as best I could.
This class looks like it was added but only half-finished.  I edited it
and documented it as best I could, and removed functions that were
unused, unimplemented, _and_ did not comply with Ifpack2 naming
conventions (e.g., "SetParameters").

I strongly encourage developers to make use of the Ifpack2::Details
namespace for classes and functions that they need in order to get their
work done, but that are not yet ready to release to the public.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=22,notpassed=0 (1.29 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.67 min)