Added a file with a very brief explanation on how to configure Trilinos
In an attempt to reduce information overload when someone is trying to configure
Trilinos for the first time we decided we needed a simple and brief explanation
that avoids most of the options available to the configure. This INSTALL file
has 4 examples on how to configure build and install Trilinos and a very brief
run down of important options. This document does direct people to the more
verbose TrilinosBuildQuickRef document if they want more information, but then
provides enough to get people moving if they want a quick build or just need a
reminder on how to configure.