Ifpack2: Remove redundant option
There was an explicitly defined Ifpack2_ENABLE_Zoltan2 option in the Ifpack2
CMakeLists.txt file that was interfering with the Tribits supplied version of
this variable. The explicitly defined variable was defaulting to the opposite of
what Tribits would when both Ifpack2 and Zoltan2 were enabled which was causing
some confusion in external projects. Removing the explicit definition so that
the Tribits supplied variable will be used.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=22,notpassed=0 (1.10 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (0.63 min)