MueLu: Amesos2?Smoother: cleaned up through query
In case user messes up, or if some external package is not available,
then we try to fall back to default.

There is one failing test:

        FactoryFactory_BuildFactory_UnitTest

I am not sure if that test makes sense any longer, as we try very hard
to find an existing direct solver which would depend on Trilinos
configuration. I have disabled this test for now (famous last words).

amend