Ifpack2: ugly hack in setParameters to deal with const ParameterList
Originally, setParameters made a non-const copy of the provided
parameter list. The problem with that is that application has no way of
knowing which provided parameters are being used as printing out
parameter list results in [unused].

Note: the behaviour is now similar to that of Chebyshev. If someone
properly fixes this, he would also need to fix Chebyshev::setParameters.