Panzer: Modifying internals of DOFManager to make ghosted elements easier
Moved some code that could be reused into projected methods. I also
added a block of code that would optionally add the ghosted gids to
the dof manager. This capability is currently turned off as it may
increase the expense of a standard continuous galerkin problem.

It remains to convert over getLocalIds