MueLu: fixed parameter handling in smoothers
When I moved smoothers to use ParameterListAcceptors, I forgot that we
pass those parameters to TPLs, possibly including factories. For
instance, we could pass ("A", *) to Ifpack2 setParameters which properly
threw an error there.
Fixed that by removing any factories from lists we pass to TPLs. Not
perfect, but works.