Tpetra: Rearranged examples.
Most of the Tpetra examples are actually for expert users and Tpetra
developers, not for novice users.  I moved these examples (that are not
for novice users) into a new example/advanced directory.  I also moved
the Epetra / Tpetra scaling test into a new example/not-built directory,
since it is not built or run by CMake.

In later commits, I will add new examples that are better suited for
novice users.