MueLu: fixing easy interpreter tests for Chebyshev
Chebyshev produces different eigenvalues depending on the compiler
version. This happens because Tpetra random vector, used as an initial
guess in Ifpack2::Chebyshev's power method is using the stdlib srand()
and rand(), which are guaranteed to behave in the same way for all
compiler versions. Epetra, on the other hand, rolls out its own random
number generator, and this allows us to produce same results independent
of the environment.
This patch introduces an extra preprocessing steps by removing the
computed Chebyshev values before comparing test and run output files.