Ifpack2: another unit test for Additive Schwarz
This test attempts to verify that transfers to and from subdomains in Additive Schwarz are correct.

It creates a global vector g populated with GIDs:  g(i) = A->getRowmap->getGlobalElement(i).
An (AS + mock subdomain solver) preconditioner is created and applied to g.  The mock subdomain
solver Ifpack2::SolverForTesting simply copies x to y, where A(i)y = x is the subdomain system to
be solved.  Any overlap can be used, with or without rcm, but the combine mode must be Tpetra::Zero.

In theory, if transfers to and from the subdomains are correct, then g'=g.