Ifpack2: Deprecated condition number estimation.
Refer to discussion of Bug 5841 for details.  Apparently nobody uses
this feature, even in Ifpack ("Ifpack1").  It never worked how people
thought it worked, in either Ifpack or Ifpack2.  I'm deprecating it so
nobody else is tempted to use it.

I also fixed the condition number estimation of Diagonal to return -1
(as it should) instead of +1 in the "not yet computed" case.