Ifpack2: LocalFilter: Start (not end) of fix for Bug 5992.
In a debug build (HAVE_IFPACK2_DEBUG), LocalFilter now checks that the
(range, row) Maps and the (domain, column) Maps are "fitted." This is a
necessary condition for implementing LocalFilter in a fast and frugal
way.  The "fitted" property is local (per MPI process), so the
optimization will be per-process as well.