Ifpack2: OverlappingRowMatrix: Construction optimization.
OverlappingRowMatrix is an implementation detail of (overlapping)
AdditiveSchwarz.  Its constructor builds the overlapping matrix's Maps
by looping over overlap levels.  This uses Import on a CrsMatrix as an
implementation detail.  (This is inefficient; see discussion of Bug
5992.)  The last iteration of the loop was doing a bunch of work and
then throwing it away.  For example, it was constructing a matrix and an
Import object, executing the Import, then calling fillComplete on the
matrix.  This commit skips that work in the last iteration of the loop.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=110,notpassed=0 (1.86 min)
1) SERIAL_RELEASE => passed: passed=88,notpassed=0 (0.14 min)