Kokkos: New Kokkos wrapper Node can now be DefaultNode.
WARNING: This commit will break Tpetra and its downstream packages.  We
have separated it out into a separate commit in order to make a nicer
commit message, and also because other files that we had to change (in
order to build tests) don't have much to do with the goal of this
commit.

NOTE: Please blame Mark Hoemmen <mhoemme@sandia.gov> (who is writing this
commit message) for any breakage that results from this commit.

The goal of this commit is to make it possible for
KokkosClassic::DefaultNode::DefaultNodeType to be any of the new Kokkos
wrapper Node types.  These types include the following:

  - Kokkos::Compat::KokkosCudaWrapperNode
  - Kokkos::Compat::KokkosOpenMPWrapperNode
  - Kokkos::Compat::KokkosThreadsWrapperNode

You may specify the default Node type using the CMake option
KokkosClassic_DefaultNode.  For example:

KokkosClassic_DefaultNode:STRING="Kokkos::Compat::KokkosOpenMPWrapperNode"

In order to make this work, we had to do the following:

1. We reversed the dependency between KokkosClassic and KokkosCompat.
It used to be that KokkosCompat had a required dependency on
KokkosClassic.  Now KokkosClassic has an optional dependency on
KokkosCompat.  We will need to move one unit test into KokkosClassic in
order to make all the tests work; we disabled that one test for now.

2. We moved the Kokkos::Compat::NodeDevice struct (which maps from
KokkosClassic Node type to new Kokkos Device type) from KokkosCompat to
KokkosClassic.

3. We fixed some use of const and static in KokkosClassic functors.
This should only be a problem for users who are writing their own
KokkosClassic functors and were depending on functor methods being
class (static) methods, instead of const instance methods.

Subsequent commits will address build issues in Kokkos and Tpetra that
this commit introduces.